Split functions to Utils file
