+++ Ramsey.py
@@ -1,6 +1,6 @@
 import copy
 import sys
-
+from RamseyUtils import *
 
 def is_bit_set(num, bit):
     return num & (1 << bit) > 0
@@ -14,17 +14,7 @@ def subsets(s):
     return sets
 
 
-def get_subset(super_set, k, idx, current):
-    if len(current) == k:
-        x = copy.deepcopy(current)
-        solution.append(x)
-        return
-    if idx == len(super_set):
-        return
-    current.append(super_set[idx])
-    get_subset(super_set, k, idx + 1, current)
-    current.remove(super_set[idx])
-    get_subset(super_set, k, idx + 1, current)
+
 
 
 def check_graph(graph, color1, color2):
